refactor(step1): request, response DTO 레이어 변경
계층 구조를 지키기위해 변경
