+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -6,8 +6,8 @@ import kitchenpos.menus.domain.PricePolicyImpl;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.tobe.domain.*;
 import kitchenpos.products.infra.PurgomalumClient;
-import kitchenpos.products.ui.dto.ProductChangePriceRequest;
-import kitchenpos.products.ui.dto.ProductCreateRequest;
+import kitchenpos.products.application.dto.ProductChangePriceRequest;
+import kitchenpos.products.application.dto.ProductCreateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
